// src/utils/types/bet-type.ts
export enum BetType {
  WINNER = "Победит",
  FIRST_OR_SECOND = "Место 1-2",
  FIRST_BETWEEN_THIRD = "Место 1-3",
  EVEN = "Чётное",
  ODD = "Нечётное",
  ABOVE = "Больше",
  BELOW = "Меньше",
  TRIFECTA = "Trifecta",
  EXACTA = "Exacta",
  QUINELLA = "Quinella",
}